+++ tox.ini
@@ -3,7 +3,7 @@ envlist=py35, flake8
 
 [testenv]
 commands=
-    python -m unittest discover './tests' '*_test.py'
+    python -m unittest discover './tests' 'test_*.py'
 deps=
     six
